# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Freelancer Weekly Planner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  name: string;
  profession: string;
  monthlyGoal?: number;
  weeklyHours: number;
  currency: string;
  niche: string;
  workDays: string[];
  /** @nullable */
  timezone?: string | null;
  weekStartDay?: string;
  darkMode?: boolean;
}

export interface ProfileInput {
  name: string;
  profession: string;
  monthlyGoal: number;
  weeklyHours: number;
  currency: string;
  niche: string;
  workDays: string[];
  timezone?: string;
  weekStartDay?: string;
  darkMode?: boolean;
}

export interface ProfileUpdate {
  name?: string;
  profession?: string;
  monthlyGoal?: number;
  weeklyHours?: number;
  currency?: string;
  niche?: string;
  workDays?: string[];
  timezone?: string;
  weekStartDay?: string;
  darkMode?: boolean;
}

export interface Client {
  id: number;
  name: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  projectType?: string | null;
  rate: number;
  currency: string;
  paymentStatus?: string;
  /** lead | active | waiting | completed | ghosted */
  status: string;
  /** @nullable */
  lastContactDate?: string | null;
  /** @nullable */
  notes?: string | null;
  totalEarned?: number;
  outstandingAmount?: number;
  projectsCompleted?: number;
  createdAt?: string;
}

export interface ClientInput {
  name: string;
  company?: string;
  email?: string;
  country?: string;
  projectType?: string;
  rate: number;
  currency: string;
  paymentStatus?: string;
  status: string;
  lastContactDate?: string;
  notes?: string;
}

export interface ClientUpdate {
  name?: string;
  company?: string;
  email?: string;
  country?: string;
  projectType?: string;
  rate?: number;
  currency?: string;
  paymentStatus?: string;
  status?: string;
  lastContactDate?: string;
  notes?: string;
}

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** ISO date YYYY-MM-DD */
  scheduledDate: string;
  /** @nullable */
  estimatedHours?: number | null;
  /** @nullable */
  actualHours?: number | null;
  /** low | medium | high | urgent */
  priority: string;
  /** todo | in_progress | waiting | delivered | revision | completed */
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
}

export interface TaskInput {
  title: string;
  clientId?: number;
  deadline?: string;
  scheduledDate: string;
  estimatedHours?: number;
  actualHours?: number;
  priority: string;
  status: string;
  notes?: string;
}

export interface TaskUpdate {
  title?: string;
  clientId?: number;
  deadline?: string;
  scheduledDate?: string;
  estimatedHours?: number;
  actualHours?: number;
  priority?: string;
  status?: string;
  notes?: string;
}

export interface IncomeEntry {
  id: number;
  amount: number;
  currency: string;
  date: string;
  description: string;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  invoiceNumber?: string | null;
  /** paid | pending | overdue */
  paymentStatus?: string;
  createdAt?: string;
}

export interface IncomeInput {
  amount: number;
  currency: string;
  date: string;
  description: string;
  clientId?: number;
  invoiceNumber?: string;
  paymentStatus?: string;
}

export interface IncomeUpdate {
  amount?: number;
  currency?: string;
  date?: string;
  description?: string;
  clientId?: number;
  invoiceNumber?: string;
  paymentStatus?: string;
}

export interface IncomeGoal {
  id: number;
  /** weekly | monthly | quarterly */
  period: string;
  targetAmount: number;
  currency: string;
  currentAmount?: number;
  periodLabel?: string;
  createdAt?: string;
}

export interface IncomeGoalInput {
  period: string;
  targetAmount: number;
  currency: string;
  periodLabel?: string;
}

export interface IncomeGoalUpdate {
  period?: string;
  targetAmount?: number;
  currency?: string;
  periodLabel?: string;
}

export interface Expense {
  id: number;
  amount: number;
  currency: string;
  date: string;
  description: string;
  /** software | ads | internet | team | equipment | misc */
  category: string;
  recurring?: boolean;
  createdAt?: string;
}

export interface ExpenseInput {
  amount: number;
  currency: string;
  date: string;
  description: string;
  category: string;
  recurring?: boolean;
}

export interface ExpenseUpdate {
  amount?: number;
  currency?: string;
  date?: string;
  description?: string;
  category?: string;
  recurring?: boolean;
}

export interface Goal {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** financial | skill | client | productivity */
  category: string;
  /** not_started | in_progress | completed */
  status: string;
  /** 0-100 percent */
  progress?: number;
  /** @nullable */
  deadline?: string | null;
  createdAt?: string;
}

export interface GoalInput {
  title: string;
  description?: string;
  category: string;
  status: string;
  progress?: number;
  deadline?: string;
}

export interface GoalUpdate {
  title?: string;
  description?: string;
  category?: string;
  status?: string;
  progress?: number;
  deadline?: string;
}

export interface DashboardSummary {
  weeklyRevenue: number;
  /** 0-100 percentage */
  monthlyGoalProgress: number;
  monthlyGoalTarget?: number;
  monthlyGoalCurrent?: number;
  pendingInvoices: number;
  activeClients: number;
  hoursWorkedThisWeek: number;
  /** 0-100 */
  productivityScore: number;
  totalTasksThisWeek: number;
  completedTasksThisWeek: number;
  netProfitThisMonth: number;
  totalExpensesThisMonth: number;
  currency?: string;
}

export interface WeeklyRevenuePoint {
  /** Label like "Jan 6" */
  week: string;
  revenue: number;
}

export interface ClientDistributionPoint {
  clientName: string;
  revenue: number;
  percentage: number;
}

export interface Insight {
  id: string;
  /** workload | income | client | productivity */
  type: string;
  message: string;
  /** info | warning | critical */
  severity: string;
}

export interface TaskCompletionPoint {
  week: string;
  completed: number;
  total: number;
  /** 0-100 percent */
  rate: number;
}

export type ListTasksParams = {
  /**
   * ISO date for the start of the week (YYYY-MM-DD)
   */
  weekStart?: string;
  clientId?: number;
  status?: string;
};

export type ListIncomeParams = {
  /**
   * YYYY-MM
   */
  month?: string;
};

export type ListExpensesParams = {
  /**
   * YYYY-MM
   */
  month?: string;
};