# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Freelancer Weekly Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: Client list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Tasks
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            description: YYYY-MM
      responses:
        "200":
          description: Expense list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Goals
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List all goals
      responses:
        "200":
          description: Goal list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Dashboard / Analytics
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, name, status, rate, currency]
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        projectType:
          type: ["string", "null"]
        rate:
          type: number
        currency:
          type: string
        paymentStatus:
          type: string
        status:
          type: string
          description: "lead | active | waiting | completed | ghosted"
        lastContactDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        totalEarned:
          type: number
        outstandingAmount:
          type: number
        projectsCompleted:
          type: integer
        createdAt:
          type: string
    ClientInput:
      type: object
      required: [name, status, rate, currency]
      properties:
        name:
          type: string
        company:
          type: string
        email:
          type: string
        country:
          type: string
        projectType:
          type: string
        rate:
          type: number
        currency:
          type: string
        paymentStatus:
          type: string
        status:
          type: string
        lastContactDate:
          type: string
        notes:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        company:
          type: string
        email:
          type: string
        country:
          type: string
        projectType:
          type: string
        rate:
          type: number
        currency:
          type: string
        paymentStatus:
          type: string
        status:
          type: string
        lastContactDate:
          type: string
        notes:
          type: string
...
    Expense:
      type: object
      required: [id, amount, currency, date, description, category]
      properties:
        id:
          type: integer
        amount:
          type: number
        currency:
          type: string
        date:
          type: string
        description:
          type: string
        category:
          type: string
          description: "software | ads | internet | team | equipment | misc"
        recurring:
          type: boolean
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [amount, currency, date, description, category]
      properties:
        amount:
          type: number
        currency:
          type: string
        date:
          type: string
        description:
          type: string
        category:
          type: string
        recurring:
          type: boolean
    ExpenseUpdate:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
        date:
          type: string
        description:
          type: string
        category:
          type: string
        recurring:
          type: boolean
    Goal:
      type: object
      required: [id, title, category, status]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          description: "financial | skill | client | productivity"
        status:
          type: string
          description: "not_started | in_progress | completed"
        progress:
          type: integer
          description: "0-100 percent"
        deadline:
          type: ["string", "null"]
        createdAt:
          type: string
    GoalInput:
      type: object
      required: [title, category, status]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
        progress:
          type: integer
        deadline:
          type: string
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
        progress:
          type: integer
        deadline:
          type: string
...