# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Freelancer Weekly Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  amount: number;
  currency: string;
  date: string;
  description: string;
  /** software | ads | internet | team | equipment | misc */
  category: string;
  recurring?: boolean;
  createdAt?: string;
}