# lib/api-zod/src/generated/types/expenseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Freelancer Weekly Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseInput {
  amount: number;
  currency: string;
  date: string;
  description: string;
  category: string;
  recurring?: boolean;
}