# lib/api-zod/src/generated/types/goalInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Freelancer Weekly Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface GoalInput {
  title: string;
  description?: string;
  category: string;
  status: string;
  progress?: number;
  deadline?: string;
}